backend/requirements.txt
python-dotenv
fastapi[standard]
openai
uvicorn
sqlmodel
psycopg[binary]
psycopg[pool]
pydantic